Move timestamp parsing methods into Util class
