+++ src/Pmi/EntityManager/EntityManager.php
@@ -1,6 +1,8 @@
 <?php
 namespace Pmi\EntityManager;
 
+use Pmi\Util;
+
 class EntityManager
 {
     protected $dbal;
@@ -46,7 +48,7 @@ class EntityManager
     public function fetchAll($query, $parameters)
     {
         $result = $this->dbal->fetchAll($query, $parameters);
-        return $this->parseMultipleTimestamps($result);
+        return Util::parseMultipleTimestamps($result, $this->timezone);
     }
 
     public function setTimezone($timezone)
@@ -57,34 +59,5 @@ class EntityManager
     public function getTimezone()
     {
         return $this->timezone;
-    }
-
-    protected function parseMultipleTimestamps(array $result)
-    {
-        foreach ($result as $key => $value) {
-            $result[$key] = $this->parseTimestamps($value);
-        }
-        return $result;
-    }
-
-    protected function parseTimestamps(array $result)
-    {
-        foreach ($result as $key => $value) {
-            if (null !== $value && substr($key, -3, 3) == '_ts' && preg_match("/^\d{4}\-\d{2}\-\d{2}/", $value)) {
-                $result[$key] = \DateTime::createFromFormat('Y-m-d H:i:s', $value)->setTimezone(new \DateTimeZone($this->timezone));
-            }
-        }
-        return $result;
-    }
-
-    protected function dateTimesToStrings(array $data)
-    {
-        foreach ($data as $key => $value) {
-            if ($value instanceof \DateTime) {
-                $value->setTimezone(new \DateTimezone('UTC'));
-                $data[$key] = $value->format('Y-m-d H:i:s');
-            }
-        }
-        return $data;
     }
 }
